requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
coverage==7.3.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
execnet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-xdist==3.3.1
python-dotenv==1.0.0
pytz==2023.3
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.12
wsproto==1.2.0
